+++ src/components/draw/KeywordSlider.jsx
